pyproject.toml
[tool.poetry]
name = "homework-7-1"
version = "0.1.0"
description = ""
authors = ["Dontgoingforyou <mishael000@yandex.ru>"]
readme = "README.md"
packages = [{include = "homework_7"}]

[tool.poetry.dependencies]
python = "^3.12"
pillow = "^11.0.0"
psycopg2-binary = "^2.9.10"
django = "^5.1.3"
djangorestframework = "^3.15.2"
flake8 = "^7.1.1"
python-dotenv = "^1.0.1"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
